+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Bump Mixpanel to v5.9.5
 - Remove post delayed callbacks when view is detached from window
 - Remove canceling previously scheduled periodic work in `SyncScheduler`
+- Change SQL performance reporting to use begin and end hooks
 
 ### Changes