+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Remove post delayed callbacks when view is detached from window
 - Remove canceling previously scheduled periodic work in `SyncScheduler`
 - Change SQL performance reporting to use begin and end hooks
+- Use view binding to inflate layout in `BaseDialog`
 
 ### Changes