+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Remove canceling previously scheduled periodic work in `SyncScheduler`
 - Change SQL performance reporting to use begin and end hooks
 - Use view binding to inflate layout in `BaseDialog`
+- Fix overdue list file name in `OverdueListDownloader`
 
 ### Changes